fix: handle errors from json.Marshal and json.Encode in handlers
- handler: add warning log when json.Encode fails in writeJSON
- middleware/workspace: handle json.Marshal error in writeError
- collaboration service: handle json.Marshal errors in StoreMemory and SaveCheckpoint
- runtime_pool: fix type mismatch in fallback token generation

+++ server/internal/handler/runtime_pool.go
@@ -621,7 +621,7 @@ func generateRandomToken(length int) string {
 		// Fall back to a less secure pseudo-random source only if crypto/rand fails.
 		// This should virtually never happen.
 		for i := range b {
-			b[i] = charset[time.Now().UnixNano()%byte(len(charset))]
+			b[i] = charset[int(time.Now().UnixNano())%len(charset)]
 		}
 		return string(b)
 	}
